requirements.txt
flask
google-generativeai
youtube-transcript-api==0.6.2
gunicorn
requests